Moving evidenceClient to aws_tests.
