+++ src/run_audit.py
@@ -1,7 +1,6 @@
 import aws_tests
 from utils import confirm_delete_folder, load_config, create_session, get_aws_account_id
 from build_report import generate_pdf_report
-from evidenceClient import EvidenceClient
 import json
 
 class Audit:
@@ -12,7 +11,7 @@ class Audit:
     config_file_path="config.json"):
         self.evidence_folder = f"{tmp_folder}/audit_evidence"       # Name of the evidence_folder
         self.config = load_config(config_file_path)                 # Control and sample exclusions
-        self.evidence_client = EvidenceClient(base_path=self.evidence_folder, debug=False)
+        self.evidence_client = aws_tests.EvidenceClient(base_path=self.evidence_folder, debug=False)
         self.session = create_session()
         self.in_scope_regions = aws_tests.get_regions(self)        
         self.aws_account_id = get_aws_account_id(self)
